+++ src/scenes/bill/components/index.js
@@ -1,2 +1,3 @@
 // @flow
+export { default as Content } from './content'
 export { Element } from './element'
